Move packer test helpers to new file
